Temp fix for #269
So the problem is because we define variables based off the name the
user gives us, and if they use a hypen or space in the challenge, it'll
cause that code to do a flip.

Until I can think of a way to detect when they pass that non-string
version of their name into `alert`, we're just going to not do that and
give them the no alert error message.